Test changesets filtering, part 1
